requirements.txt
fastapi
uvicorn
scikit-learn
joblib
numpy
kafka-python
requests
